Switch start-canton.sh to spin up Canton against Postgres (#604)
* Use Postgres for tests

* Address shellcheck warnings

* Support both docker and local postgres

* Fix after rebase

* Maybe fix CI

* Maybe fix CI

* Maybe fix CI

* Maybe fix CI

* Maybe fix CI

* Maybe fix CI

* Maybe fix CI

* Maybe fix CI

* Remove test flake

* Remove test flake

* format

* Maybe fix CI

* Remove test flake

+++ stop-canton.sh
@@ -1,6 +1,14 @@
 #!/usr/bin/env bash
 set -eou pipefail
 
-PID=$(cat canton.pid)
-kill $PID
-rm canton.pid
+POSTGRES_MODE=${1:-docker}
+
+if [ -f "canton.pid" ]; then
+    PID=$(cat canton.pid)
+    kill "$PID"
+    rm canton.pid
+else
+    echo "The file canton.pid does not exist, not stopping Canton"
+fi
+
+./scripts/postgres.sh "$POSTGRES_MODE" stop
